apps/pptx_generator/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-pptx>=0.6.23
pillow>=10.2.0
pandas>=2.2.0
openpyxl>=3.1.2
pyyaml>=6.0.1
aiofiles>=23.2.1
httpx>=0.26.0
matplotlib>=3.8.0
numpy>=1.26.0
ruff>=0.1.14
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
mypy>=1.8.0
